Clear the three build warnings, none of them reachable
Release prep asks for a 0-warning Release build and the tree was at
three, all cosmetic, all traced before touching:

CS0108 x2, the Edge's AppearanceIds/AppearanceNames shadowing the
DualSense arrays. Statics cannot be virtual, so this is name shadowing
and `new` states the intent that was already in the file's header
comment. Nothing resolves them through a base-typed expression:
ControllerModelView's family switch names the Edge type explicitly, and
the Edge reaches the shared body through the protected family-scoped
constructor, so the base's private Validate never sits on its path.

CS0675, a sign-extended int widening to ulong in the Bluetooth address
parser. Every char reaching that line passed IsHex, so the nibble is
0..15 and the sign bit can never be set. Widening through uint says so
explicitly.

Why they survived several audit rounds: a delta-scoped audit correctly
files a pre-existing warning as out of scope, and does that every round,
so the baseline only ever ratchets up. Round 43 recorded all three and
adjudicated them exactly that way.

+++ PadForge.App/Services/HeadsetTrackerRepair.cs
@@ -273,7 +273,12 @@ namespace PadForge.Services
                 for (int i = 0; i < 12; i++)
                 {
                     char c = text[pos + i];
-                    value = (value << 4) | (ulong)(c <= '9' ? c - '0' : c - 'A' + 10);
+                    // (uint) before (ulong): the ternary is an int, and a
+                    // direct int->ulong cast sign-extends. Every char here
+                    // passed IsHex, so the nibble is 0..15 and can never be
+                    // negative, but widening through uint states that
+                    // instead of leaving the compiler to warn about it.
+                    value = (value << 4) | (uint)(c <= '9' ? c - '0' : c - 'A' + 10);
                 }
                 return value != 0;
             }
